refactor: analytics context event
