+++ src/components/common/Extensions/UserHubButton/UserHubButton.tsx
@@ -15,15 +15,10 @@ import Button from '~v5/shared/Button';
 import UserAvatar from '~v5/shared/UserAvatar';
 import PopoverBase from '~v5/shared/PopoverBase';
 import useNavigationSidebarContext from '~v5/frame/NavigationSidebar/partials/NavigationSidebarContext/hooks';
-import {
-  AnalyticsEventAction,
-  AnalyticsEventCategory,
-  AnalyticsEventLabel,
-  AnalyticsEventType,
-  useAnalyticsContext,
-} from '~context/AnalyticsContext';
+import { useAnalyticsContext } from '~context/AnalyticsContext';
 
 import { UserHubButtonProps } from './types';
+import { OPEN_USER_HUB_EVENT } from './consts';
 
 export const displayName =
   'common.Extensions.UserNavigation.partials.UserHubButton';
@@ -43,14 +38,6 @@ const UserHubButton: FC<UserHubButtonProps> = ({
   const { setOpenItemIndex, mobileMenuToggle } = useNavigationSidebarContext();
 
   const [, { toggleOff }] = mobileMenuToggle;
-  const trackUserHubClick = () => {
-    trackEvent(
-      AnalyticsEventType.CUSTOM_EVENT,
-      AnalyticsEventCategory.USER,
-      AnalyticsEventAction.CLICK,
-      AnalyticsEventLabel.OPEN_USER_HUB,
-    );
-  };
 
   const popperTooltipOffset = isMobile ? [0, 1] : [0, 8];
 
@@ -89,8 +76,9 @@ const UserHubButton: FC<UserHubButtonProps> = ({
     );
 
   useDisableBodyScroll(visible && isMobile);
+
   const handleButtonClick = () => {
-    trackUserHubClick();
+    trackEvent(OPEN_USER_HUB_EVENT);
     setOpenItemIndex(undefined);
     toggleOff();
   };
