+++ src/components/common/Extensions/UserHubButton/consts.ts
@@ -1,0 +1,14 @@
+import {
+  AnalyticsEvent,
+  AnalyticsEventAction,
+  AnalyticsEventCategory,
+  AnalyticsEventLabel,
+  AnalyticsEventType,
+} from '~context/AnalyticsContext';
+
+export const OPEN_USER_HUB_EVENT: AnalyticsEvent = {
+  event: AnalyticsEventType.CUSTOM_EVENT,
+  category: AnalyticsEventCategory.USER,
+  action: AnalyticsEventAction.CLICK,
+  label: AnalyticsEventLabel.OPEN_USER_HUB,
+};
