+++ src/context/ActionSidebarContext.tsx
@@ -22,6 +22,7 @@ import {
   AnalyticsEventCategory,
   AnalyticsEventAction,
   AnalyticsEventLabel,
+  AnalyticsEvent,
 } from './AnalyticsContext';
 
 type ActionSidebarToggle = [
@@ -45,6 +46,13 @@ export const ActionSidebarContext = createContext<ActionSidebarContextValue>({
   actionSidebarToggle: DEFAULT_USE_TOGGLE_RETURN_VALUE,
   cancelModalToggle: DEFAULT_USE_TOGGLE_RETURN_VALUE,
 });
+
+const OPEN_ACTION_PANEL_EVENT: AnalyticsEvent = {
+  event: AnalyticsEventType.CUSTOM_EVENT,
+  category: AnalyticsEventCategory.ACTION_PANEL,
+  action: AnalyticsEventAction.TRIGGER,
+  label: AnalyticsEventLabel.OPEN_ACTION_PANEL,
+};
 
 export const ActionSidebarContextProvider: FC<PropsWithChildren> = ({
   children,
@@ -91,12 +99,7 @@ export const ActionSidebarContextProvider: FC<PropsWithChildren> = ({
     (initialValues) => {
       setActionSidebarInitialValues(initialValues);
       // Track the event when the action panel is opened
-      trackEvent(
-        AnalyticsEventType.CUSTOM_EVENT,
-        AnalyticsEventCategory.ACTION_PANEL,
-        AnalyticsEventAction.TRIGGER,
-        AnalyticsEventLabel.OPEN_ACTION_PANEL,
-      );
+      trackEvent(OPEN_ACTION_PANEL_EVENT);
       return toggleActionSidebarOn();
     },
     [toggleActionSidebarOn, trackEvent],
