+++ src/context/AnalyticsContext/AnalyticsContext.tsx
@@ -8,62 +8,10 @@ import React, {
   useState,
 } from 'react';
 import { useBeamer } from '~hooks/useBeamer';
-
-export enum AnalyticsEventType {
-  CUSTOM_EVENT = 'custom_event',
-}
-
-export enum AnalyticsEventCategory {
-  USER = 'user',
-  ACTION_PANEL = 'action_panel',
-  ADMIN = 'admin',
-}
-
-export enum AnalyticsEventAction {
-  CLICK = 'CLICK',
-  TRIGGER = 'TRIGGER',
-  TRANSACTION = 'TRANSACTION',
-}
-
-export enum AnalyticsEventLabel {
-  OPEN_USER_HUB = 'Open user hub',
-  OPEN_ACTION_PANEL = 'Open action panel',
-}
-
-// Analytics event
-export interface AnalyticsEvent {
-  event: AnalyticsEventType;
-  category: AnalyticsEventCategory;
-  action: AnalyticsEventAction;
-  label?: AnalyticsEventLabel;
-  value?: number;
-}
-
-// Helper function to create an analytics event
-const createAnalyticsEvent = (
-  eventType: AnalyticsEventType,
-  eventCategory: AnalyticsEventCategory,
-  eventAction: AnalyticsEventAction,
-  eventLabel: AnalyticsEventLabel,
-  value?: number,
-): AnalyticsEvent => {
-  return {
-    event: eventType,
-    category: eventCategory,
-    action: eventAction,
-    label: eventLabel,
-    value,
-  };
-};
+import { AnalyticsEvent } from './types';
 
 interface AnalyticsContextValue {
-  trackEvent: (
-    eventType: AnalyticsEventType,
-    eventCategory: AnalyticsEventCategory,
-    eventAction: AnalyticsEventAction,
-    eventLabel: AnalyticsEventLabel,
-    value?: number,
-  ) => void;
+  trackEvent: (event: AnalyticsEvent) => void;
   trackPageView: (path: string) => void;
 }
 
@@ -106,25 +54,12 @@ export const AnalyticsContextProvider: FC<PropsWithChildren> = ({
   const contextValue = useMemo(() => {
     return isAnalyticsActive
       ? {
-          trackEvent: (
-            eventType,
-            eventCategory,
-            eventAction,
-            eventLabel,
-            value,
-          ) => {
-            const event = createAnalyticsEvent(
-              eventType,
-              eventCategory,
-              eventAction,
-              eventLabel,
-              value,
-            );
+          trackEvent: (event: AnalyticsEvent) => {
             if (window.dataLayer) {
               window.dataLayer.push(event);
             }
           },
-          trackPageView: (path) => {
+          trackPageView: (path: string) => {
             if (window.dataLayer) {
               window.dataLayer.push({ event: 'pageview', path });
             }
